Feature: live update
Calculations are moved to the Summary page, and any updates on child pages are
forwarded to the Summary Page (using a callback from the Summary Page).  Child
pages update their content whenever something changes, and the Summary Page
updates its content whenever something changes on it or any of the child pages.